+++ packages/storefront/tests/unit/assets.package.mock.js
@@ -1,9 +1,9 @@
-const METAICONS_CDN_BASE_URL = '';
-const METAICONS_MANIFEST = {
+const META_ICONS_CDN_BASE_URL = '';
+const META_ICONS_MANIFEST = {
   favicon: {},
   touchicon: {},
   pinnedTab: {},
   mstile: {}
 };
 
-module.exports = { METAICONS_CDN_BASE_URL, METAICONS_MANIFEST };
+module.exports = { META_ICONS_CDN_BASE_URL, META_ICONS_MANIFEST };