+++ packages/storefront/vue.config.js
@@ -1,9 +1,9 @@
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-const { METAICONS_MANIFEST, METAICONS_CDN_BASE_URL }  =
+const { META_ICONS_MANIFEST, META_ICONS_CDN_BASE_URL } =
   process.env.NODE_ENV !== 'test'
     ? require('@porsche-design-system/assets')
     : require('./tests/unit/assets.package.mock.js');
-const { favicon, touchicon, pinnedTab, mstile } = METAICONS_MANIFEST;
+const { favicon, touchicon, pinnedTab, mstile } = META_ICONS_MANIFEST;
 
 module.exports = {
   publicPath: './',
@@ -42,23 +42,26 @@ module.exports = {
     appleMobileWebAppCapable: 'yes',
     appleMobileWebAppStatusBarStyle: 'white',
     iconPaths: {
-      favicon16: `${METAICONS_CDN_BASE_URL}/${favicon.favicon_16x16}`,
-      favicon32: `${METAICONS_CDN_BASE_URL}/${favicon.favicon_32x32}`,
-      favicon48: `${METAICONS_CDN_BASE_URL}/${favicon.favicon_48x48}`,
-      appleTouchIcon: `${METAICONS_CDN_BASE_URL}/${touchicon.appleTouchIcon_180x180}`,
-      maskIcon: `${METAICONS_CDN_BASE_URL}/${pinnedTab.pinnedTabIcon}`,
-      msTileImage: `${METAICONS_CDN_BASE_URL}/${mstile.mstile_270x270}`,
+      favicon16: `${META_ICONS_CDN_BASE_URL}/${favicon.favicon_16x16}`,
+      favicon32: `${META_ICONS_CDN_BASE_URL}/${favicon.favicon_32x32}`,
+      favicon48: `${META_ICONS_CDN_BASE_URL}/${favicon.favicon_48x48}`,
+      appleTouchIcon: `${META_ICONS_CDN_BASE_URL}/${touchicon.appleTouchIcon_180x180}`,
+      maskIcon: `${META_ICONS_CDN_BASE_URL}/${pinnedTab.pinnedTabIcon}`,
+      msTileImage: `${META_ICONS_CDN_BASE_URL}/${mstile.mstile_270x270}`
     },
     manifestOptions: {
-      icons: [{
-        'src': `${METAICONS_CDN_BASE_URL}/${touchicon.androidChrome_192x192}`,
-        'sizes': '192x192',
-        'type': 'image/png'
-      }, {
-        'src': `${METAICONS_CDN_BASE_URL}/${touchicon.androidChrome_512x512}`,
-        'sizes': '512x512',
-        'type': 'image/png'
-      }],
+      icons: [
+        {
+          src: `${META_ICONS_CDN_BASE_URL}/${touchicon.androidChrome_192x192}`,
+          sizes: '192x192',
+          type: 'image/png'
+        },
+        {
+          src: `${META_ICONS_CDN_BASE_URL}/${touchicon.androidChrome_512x512}`,
+          sizes: '512x512',
+          type: 'image/png'
+        }
+      ],
       background_color: '#FFFFFF'
     }
   }