+++ packages/assets/CHANGELOG.md
@@ -30,7 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### [1.1.0] - 2020-08-07
 
 #### Added
-- Metaicons manifest and CDN URLs
+- Meta Icons manifest and CDN URLs
 
 ### [1.0.0] - 2020-07-17