+++ packages/assets/CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [Unreleased]
 
+#### Changed
+- Rename `METAICONS_CDN_BASE_URL` to `META_ICONS_CDN_BASE_URL`
+- Rename `METAICONS_MANIFEST` to `META_ICONS_MANIFEST`
+
 ### [2.0.0] - 2020-08-19
 
 ### [2.0.0-rc.0] - 2020-08-17